# umc-10th-doldol2/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"name",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"startedAt": {
						"type": "string",
						"format": "date-time"
					},
					"completedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"status",
					"startedAt",
					"completedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserMissionResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionChallengeRequest": {
				"properties": {
					"missionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InProgressMissionItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"startedAt": {
						"type": "string",
						"format": "date-time"
					},
					"completedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"mission": {
						"properties": {
							"store": {
								"properties": {
									"name": {
										"type": "string"
									},
									"id": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"name",
									"id"
								],
								"type": "object"
							},
							"deadline": {
								"type": "string",
								"format": "date-time",
								"nullable": true
							},
							"reward": {
								"type": "number",
								"format": "double"
							},
							"content": {
								"type": "string"
							},
							"title": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"store",
							"deadline",
							"reward",
							"content",
							"title",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"status",
					"startedAt",
					"completedAt",
					"mission"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InProgressMissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/InProgressMissionItem"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_InProgressMissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/InProgressMissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "4week",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/users/{userId}/missions": {
			"post": {
				"operationId": "HandleMissionChallenge",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MissionChallengeRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleListInProgressMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_InProgressMissionListResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/missions/{missionId}": {
			"patch": {
				"operationId": "HandleCompleteMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/guest": {
			"get": {
				"operationId": "HandleGuestPage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/login": {
			"get": {
				"operationId": "HandleLoginPage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/mypage": {
			"get": {
				"operationId": "HandleMypage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/set-login": {
			"get": {
				"operationId": "HandleSetLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
		"/users/set-logout": {
			"get": {
				"operationId": "HandleSetLogout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}